requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
python-telegram-bot==20.0a6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
